split snapshot tests into their own file
